rename filename from constraint to criteria
